+++ src/main/kotlin/com/calmwolfs/bedwar/commands/Commands.kt
@@ -12,6 +12,7 @@ import com.calmwolfs.bedwar.features.session.SessionDisplay
 import com.calmwolfs.bedwar.utils.BedwarsStarUtils
 import com.calmwolfs.bedwar.utils.PartyUtils
 import net.minecraft.command.ICommandSender
+import net.minecraft.util.BlockPos
 import net.minecraftforge.client.ClientCommandHandler
 
 object Commands {
@@ -57,6 +58,23 @@ object Commands {
 
     private fun registerCommand(name: String, function: (Array<String>) -> Unit) {
         ClientCommandHandler.instance.registerCommand(SimpleCommand(name, createCommand(function)))
+    }
+
+    private fun registerCommand0(
+        name: String,
+        function: (Array<String>) -> Unit,
+        autoComplete: ((Array<String>) -> List<String>) = { listOf() }
+    ) {
+        val command = SimpleCommand(
+            name,
+            createCommand(function),
+            object : SimpleCommand.TabCompleteRunnable {
+                override fun tabComplete(sender: ICommandSender?, args: Array<String>?, pos: BlockPos?): List<String> {
+                    return autoComplete(args ?: emptyArray())
+                }
+            }
+        )
+        ClientCommandHandler.instance.registerCommand(command)
     }
 
     private fun createCommand(function: (Array<String>) -> Unit) = object : SimpleCommand.ProcessCommandRunnable() {